+++ src/atm_phy_aes/mo_interface_aes_tmx.f90
@@ -497,7 +497,8 @@ CONTAINS
     USE mo_nonhydro_state,     ONLY: p_nh_state
     USE mo_nonhydro_types,     ONLY: t_nh_metrics, t_nh_diag
     USE mo_dynamics_config,    ONLY: nnow, nnow_rcf
-    USE mo_physical_constants, ONLY: cpd, cpv, cvd, cvv, Tf, tmelt
+    USE mo_physical_constants, ONLY: cpd, cpv, cvd, cvv, tmelt
+    USE mo_sea_ice_nml,        ONLY: Tf
 
     USE mo_master_config, ONLY: isRestart
     USE mo_run_config,    ONLY: lmemman