+++ src/ocean/dynamics/mo_ocean_ab_timestepping_zstar.f90
@@ -53,7 +53,7 @@ MODULE mo_ocean_ab_timestepping_zstar
     & timer_extra2, timer_extra3, timer_extra4, timer_ab_expl, timer_ab_rhs4sfc, timer_total
 
   USE mo_dynamics_config,           ONLY: nold, nnew
-  USE mo_physical_constants,        ONLY: grav, clw, rho_ref, Tf
+  USE mo_physical_constants,        ONLY: grav, clw, rho_ref
   USE mo_ocean_initialization,      ONLY: is_initial_timestep
   USE mo_ocean_types, ONLY: t_hydro_ocean_state
   USE mo_ocean_time_events,   ONLY: ocean_time_nextStep, isCheckpoint, isEndOfThisRun, newNullDatetime